Currency app completed

+++ DovizUygulaması/EntityLayer/EntityKasa.cs
@@ -8,15 +8,16 @@ namespace EntityLayer
 {
     public class EntityKasa
     {
-        private int dolar;
-        private int euro;
-        private int pound;
-        private int japonyen;
-        private int turkLıra;
-        public int Dolar { get => dolar; set => dolar = value; }
-        public int Euro { get => euro; set => euro = value; }
-        public int Pound { get => pound; set => pound = value; }
-        public int Japonyen { get => japonyen; set => japonyen = value; }
-        public int TurkLıra { get => turkLıra; set => turkLıra = value; }
+        private double dolar;
+        private double euro;
+        private double pound;
+        private double japonyen;
+        private double turkLıra;
+
+        public double Dolar { get => dolar; set => dolar = value; }
+        public double Euro { get => euro; set => euro = value; }
+        public double Pound { get => pound; set => pound = value; }
+        public double Japonyen { get => japonyen; set => japonyen = value; }
+        public double TurkLıra { get => turkLıra; set => turkLıra = value; }
     }
 }
